list_models.py
import google.generativeai as genai
import os
from dotenv import load_dotenv

load_dotenv()

genai.configure(api_key=os.getenv("AIzaSyCId5GPcvi30LCOW7We7U0wVxIBJS53hMo"))

print("Available Gemini Models:\n")
print("-" * 80)

for model in genai.list_models():
    if 'generateContent' in model.supported_generation_methods:
        print(f"Model: {model.name}")
        print(f"  Display Name: {model.display_name}")
        print(f"  Description: {model.description}")
        print(f"  Supported: {model.supported_generation_methods}")
        print("-" * 80)