requirements.txt
Flask
psycopg2-binary
google-generativeai
python-dotenv
edge-tts
pydub
pypdf
python-pptx
moviepy
Pillow
textwrap3
gTTS
whispering
ffmpeg-python  
numpy          
werkzeug==2.3.7  
pdfminer.six
tabulate
langchain